+++ src/parser/inline/reference.test.ts
@@ -41,6 +41,13 @@ describe('Unit: parser/inline/reference', () => {
       assert.deepStrictEqual(inspect(parser('[[[[a]]]]')), [['<sup class="reference">[[a]]</sup>'], '']);
     });
 
+    it('alias', () => {
+      assert.deepStrictEqual(inspect(parser('[[~Xyz2020: a]]')), [['<sup class="reference" data-alias="Xyz2020">a</sup>'], '']);
+      assert.deepStrictEqual(inspect(parser('[[~Xyz2020]]')), [['<sup class="reference" data-alias="Xyz2020"></sup>'], '']);
+      assert.deepStrictEqual(inspect(parser('[[~X, 2020]]')), [['<sup class="reference" data-alias="X, 2020"></sup>'], '']);
+      assert.deepStrictEqual(inspect(parser('[[~X, Y, 2020, p1-2]]')), [['<sup class="reference" data-alias="X, Y, 2020, p1-2"></sup>'], '']);
+    });
+
   });
 
 });