+++ src/parser/inline/reference.ts
@@ -1,5 +1,5 @@
 import { ReferenceParser } from '../inline';
-import { union, some, subline, verify, surround, guard, configure, lazy, fmap } from '../../combinator';
+import { subsequence, some, subline, verify, focus, surround, guard, configure, lazy, fmap } from '../../combinator';
 import { inline } from '../inline';
 import { defrag, trimNodeEnd, hasTightText } from '../util';
 import { html } from 'typed-dom';
@@ -7,6 +7,19 @@ import { html } from 'typed-dom';
 export const reference: ReferenceParser = lazy(() => subline(verify(fmap(
   guard(config => config.syntax?.inline?.reference ?? true,
   configure({ syntax: { inline: { annotation: false, reference: false, media: false } } },
-  surround('[[', trimNodeEnd(defrag(some(union([inline]), /^\\?\n|^]]/))), ']]'))),
-  ns => [html('sup', { class: 'reference' }, ns)]),
-  ([el]) => hasTightText(el))));
+  surround('[[', subsequence([alias, trimNodeEnd(defrag(some(inline, /^\\?\n|^]]/)))]), ']]'))),
+  ns => [
+    html('sup',
+      {
+        class: 'reference',
+        'data-alias': ns[0].nodeName === 'ABBR'
+          ? ns.shift()!.textContent
+          : undefined
+      },
+      ns)
+  ]),
+  ([el]) => hasTightText(el) || el.hasAttribute('data-alias'))));
+
+const alias: ReferenceParser.AliasParser = subline(focus(
+  /^~[A-za-z][A-Za-z0-9',. -]*(?:: |(?=]]))/,
+  source => [[html('abbr', source.slice(1, source[source.length - 2] === ':' ? -2 : source.length))], '']));